feat: add new interfaces for dependency injection
This commit introduces several new interfaces in the
`OnForkHub.Core.Interfaces.DependencyInjection` namespace,
including `IAssemblyScanner`, `ILifetimeConfigurator`,
`IRegistrationResult`, and `IRegistrationStrategy`.
These interfaces provide methods for service registration
and management in a dependency injection container.
Additionally, the `ITypeSelector` interface has been
enhanced with new methods for adding classes based on
attributes and naming patterns, supporting open generics.
