+++ src/Core/OnForkHub.Core/Interfaces/DependencyInjection/IRegistrationResult.cs
@@ -1,0 +1,10 @@
+namespace OnForkHub.Core.Interfaces.DependencyInjection;
+
+public interface IRegistrationResult
+{
+    TimeSpan ElapsedTime { get; }
+
+    IReadOnlyList<Type> RegisteredTypes { get; }
+
+    int RegisteredTypesCount { get; }
+}
